linksharing: centralize error handling
We had a number of places where we were being inconsistent with
error handling in this codebase. In some places bad input would
result in a 400, in some places it would result in a 500. In
some places we'd get an HTML error page, in some places we'd get
plain text, etc. We also only logged things with a specific
action field sometimes.

This change moves all error handling up the stack, so that
handling errors is natural. If you want to indicate to the
centralized error handling in ServeHTTP what status code should
be returned, you can attach it to the error with WithStatus.

This makes it so we're more likely to be consistent with errors.

Change-Id: Ic9c330c80be076258b2ee02b9618d280988b02b2