linksharing: load index.html in parallel to reduce loading time
this reduces the loading time when pages such as 'prefix/' are
served with the expectation of returning an index.html. the only
time this turns into a throwaway, extra request with additional
work is when a user genuinely means to share an object with a
key that ends with a slash (should be very rare).

Change-Id: I073364a21e417073a0c1130c26a56a0971fc30ba