sharing: Forward client IP to Authservice
Forward the client IP to the Auth service because it will require it the
near future for mitigating malicious clients to overwhelm its DB using
correct formatted access key IDs which doesn't resolve to any existing
access grant.

Add 2 new command line flags/configurations to specify if the client IP
should be taken from HTTP headers ('Forwarded', 'X-Forwarded-For', or
'X-Real-Is') or not and allows to specify a list of IPs that are trusted
to read the headers.

Change-Id: I810766833f1231a5abee27ba4634a0f3d5d4cadd